# apps/control-plane/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenCrane Control Plane API",
    "version": "1.0.0",
    "description": "Multi-tenant AI agent platform management API. All resources require bearer-token authentication unless noted."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Versioned API prefix"
    }
  ],
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error",
          "code"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error description."
          },
          "code": {
            "type": "string",
            "description": "Machine-readable error code."
          },
          "detail": {
            "type": "string",
            "description": "Optional extra context."
          }
        }
      },
...
      "McpServer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "endpoint": {
            "type": "string"
          },
          "transport": {
            "type": "string",
            "enum": [
              "streamable-http",
              "sse",
              "websocket"
            ]
          },
          "grants": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "credentials": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "Group": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "memberCount": {
            "type": "integer"
          },
          "awarenessGrants": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
      "AccessToken": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "owner": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastUsedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Static bearer token. Pass as Authorization: Bearer <token>."
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/mcp-servers": {
      "get": {
        "operationId": "listMcpServers",
        "summary": "List all MCP servers with grants and credentials",
        "tags": [
          "MCP Servers"
        ],
        "responses": {
          "200": {
            "description": "MCP server list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/McpServer"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createMcpServer",
        "summary": "Create a new MCP server",
        "tags": [
          "MCP Servers"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "endpoint",
                  "transport"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "endpoint": {
                    "type": "string"
                  },
                  "transport": {
                    "type": "string"
                  },
                  "grants": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  },
                  "credentials": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "MCP server created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServer"
                }
              }
            }
          }
        }
      }
    },
    "/mcp-servers/{id}": {
      "get": {
        "operationId": "getMcpServer",
        "summary": "Get a single MCP server by identifier",
        "tags": [
          "MCP Servers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "MCP server detail.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServer"
                }
              }
            }
          },
          "404": {
            "description": "MCP server not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateMcpServer",
        "summary": "Update an MCP server and fully replace grants and credentials",
        "tags": [
          "MCP Servers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MCP server updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServer"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteMcpServer",
        "summary": "Delete an MCP server and its linked grant rows",
        "tags": [
          "MCP Servers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "MCP server deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/groups": {
      "get": {
        "operationId": "listGroups",
        "summary": "List all groups with member counts and awareness grants",
        "tags": [
          "Groups"
        ],
        "responses": {
          "200": {
            "description": "Group list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Group"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createGroup",
        "summary": "Create a new group and optional awareness grants",
        "tags": [
          "Groups"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Group created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          }
        }
      }
    },
    "/groups/{id}": {
      "get": {
        "operationId": "getGroup",
        "summary": "Get a single group by identifier",
        "tags": [
          "Groups"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group detail.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          },
          "404": {
            "description": "Group not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateGroup",
        "summary": "Update a group and replace awareness grants",
        "tags": [
          "Groups"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Group updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteGroup",
        "summary": "Delete a group and its awareness grants",
        "tags": [
          "Groups"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/access-tokens": {
      "get": {
        "operationId": "listAccessTokens",
        "summary": "List all issued access tokens (hashes only, never plaintext)",
        "tags": [
          "Access Tokens"
        ],
        "responses": {
          "200": {
            "description": "Token list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccessToken"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createAccessToken",
        "summary": "Create a new access token. Returns plaintext token once — store it securely.",
        "tags": [
          "Access Tokens"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "owner": {
                    "type": "string"
                  },
                  "expiresAt": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Token created. The plainTextToken field will not be returned again.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "plainTextToken"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "plainTextToken": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/access-tokens/{id}": {
      "delete": {
        "operationId": "deleteAccessToken",
        "summary": "Revoke and delete an access token",
        "tags": [
          "Access Tokens"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Token deleted."
          },
          "404": {
            "description": "Token not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}